requirements.txt
##python -m venv --system-site-packages venv
django==4.1.3
djangorestframework==3.14.0
django-cors-headers==2.2.0
pillow==9.3.3
psycopg2-binary==2.9.5